+++ app.js
@@ -3,8 +3,8 @@
 require('dotenv/config');
 
 // ℹ️ Connects to the database
-require('./db');
-
+const { mongooseConnect } = require('./db/index'); // db/index.js
+mongooseConnect();
 // Handles http requests (express is node js framework)
 // https://www.npmjs.com/package/express
 const express = require('express');
