+++ routes/index.js
@@ -1,8 +1,10 @@
 // Require express and Router to user routes
 const express = require('express');
 const router = express.Router();
+const Movie = require('../models/Movie.model');
 
 /* GET home page */
 router.get('/', (req, res, next) => res.render('index'));
+
 
 module.exports = router;
