+++ seeds/movies.seed.js
@@ -1,107 +1,64 @@
-// To insert in "seeds/movies.seed.js"
-const mongoose = require("mongoose");
-const Movie = require("../models/Movie.model");
+const {
+  mongooseClose,
+  mongooseConnect,
+  mongooseDropCollection,
+} = require("../db/index"); // db/index.js
+const Movie = require("../models/Movie.model"); // Model
+const moviesData = require("./movies.data"); // seeds/movies.data.js
 
-const MONGO_URI = process.env.MONGODB_URI || "mongodb://localhost/express-cinema";
-
-mongoose.connect(MONGO_URI, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
-  //useFindAndModify: false,
-  //useCreateIndex: true
-});
-
-
-// PASTE HERE THE LIST OF MOVIES PROVIDED
-const movies = [
-  {
-    title: "A Wrinkle in Time",
-    director: "Ava DuVernay",
-    stars: ["Storm Reid", "Oprah Winfrey", "Reese Witherspoon"],
-    image:
-      "https://images-na.ssl-images-amazon.com/images/M/MV5BMjMxNjQ5MTI3MV5BMl5BanBnXkFtZTgwMjQ2MTAyNDM@._V1_UX182_CR0,0,182,268_AL_.jpg",
-    description:
-      "Following the discovery of a new form of space travel as well as Meg's father's disappearance, she, her brother, and her friend must join three magical beings - Mrs. Whatsit, Mrs. Who, and Mrs. Which - to travel across the universe to rescue him from a terrible evil.",
-    showtimes: ["13:00", "15:30", "18:00", "20:10", "22:40"]
-  },
-  {
-    title: "The Strangers: Prey at Night",
-    director: "Johannes Roberts",
-    stars: ["Christina Hendricks", "Bailee Madison", "Martin Henderson"],
-    image:
-      "https://images-na.ssl-images-amazon.com/images/M/MV5BMTY1OTIwODgzMV5BMl5BanBnXkFtZTgwMzUyMDgzNDM@._V1_UX182_CR0,0,182,268_AL_.jpg",
-    description:
-      "A family's road trip takes a dangerous turn when they arrive at a secluded mobile home park to stay with some relatives and find it mysteriously deserted. Under the cover of darkness, three masked psychopaths pay them a visit to test the family's every limit as they struggle to survive.",
-    showtimes: ["13:50", "16:20", "19:20", "22:10"]
-  },
-  {
-    title: "The Hurricane Heist",
-    director: "Rob Cohen",
-    stars: ["Toby Kebbell", "Maggie Grace", "Ryan Kwanten"],
-    image:
-      "https://images-na.ssl-images-amazon.com/images/M/MV5BMzg3Y2MyNjgtMzk4ZS00OTU3LWEwZmMtN2Y0NTdlZjU0NGFiXkEyXkFqcGdeQXVyMTMxODk2OTU@._V1_UX182_CR0,0,182,268_AL_.jpg",
-    description:
-      "Thieves attempt a massive heist against the U.S. Treasury as a Category 5 hurricane approaches one of its Mint facilities.",
-    showtimes: ["13:10", "15:40", "18:20", "20:50", "23:10"]
-  },
-  {
-    title: "Gringo",
-    director: "Nash Edgerton",
-    stars: ["Joel Edgerton", "Charlize Theron", "David Oyelowo"],
-    image:
-      "https://images-na.ssl-images-amazon.com/images/M/MV5BMjAyMTk2MTQ3Ml5BMl5BanBnXkFtZTgwNDQ2ODE0NDM@._V1_UX182_CR0,0,182,268_AL_.jpg",
-    description:
-      "GRINGO, a dark comedy mixed with white-knuckle action and dramatic intrigue, explores the battle of survival for businessman Harold Soyinka (David Oyelowo) when he finds himself crossing the line from law-abiding citizen to wanted criminal.",
-    showtimes: ["13:40", "15:50", "19:00", "21:20", "23:50"]
-  },
-  {
-    title: "Thoroughbreds",
-    director: "Cory Finley",
-    stars: ["Anya Taylor-Joy", "Olivia Cooke", "Anton Yelchin"],
-    image:
-      "https://images-na.ssl-images-amazon.com/images/M/MV5BNDcyNDA4NDAzN15BMl5BanBnXkFtZTgwODQxMDQ5NDM@._V1_UX182_CR0,0,182,268_AL_.jpg",
-    description:
-      "Two upper-class teenage girls in suburban Connecticut rekindle their unlikely friendship after years of growing apart. Together, they hatch a plan to solve both of their problems-no matter what the cost.",
-    showtimes: ["13:00", "15:30", "18:00", "20:10", "22:40"]
-  },
-  {
-    title: "The Leisure Seeker",
-    director: "Paolo Virzì",
-    stars: ["Helen Mirren", "Donald Sutherland", "Janel Moloney"],
-    image:
-      "https://images-na.ssl-images-amazon.com/images/M/MV5BMTg1NTg2MzcyNF5BMl5BanBnXkFtZTgwNjMwMDIzNDM@._V1_UX182_CR0,0,182,268_AL_.jpg",
-    description:
-      "A runaway couple goes on an unforgettable journey in the faithful old RV they call The Leisure Seeker, traveling from Boston to The Ernest Hemingway Home in Key West. They recapture their passion for life and their love for each other on a road trip that provides revelation and surprise right up to the very end.",
-    showtimes: ["13:40", "15:50", "19:00", "21:20", "23:50"]
-  },
-  {
-    title: "Black Panther",
-    director: "Ryan Coogler",
-    stars: ["Chadwick Boseman", "Michael B. Jordan", "Lupita Nyong'o"],
-    image:
-      "https://images-na.ssl-images-amazon.com/images/M/MV5BMTg1MTY2MjYzNV5BMl5BanBnXkFtZTgwMTc4NTMwNDI@._V1_UX182_CR0,0,182,268_AL_.jpg",
-    description:
-      "T'Challa, the King of Wakanda, rises to the throne in the isolated, technologically advanced African nation, but his claim is challenged by a vengeful outsider who was a childhood victim of T'Challa's father's mistake.",
-    showtimes: ["13:50", "16:20", "19:20", "22:10"]
-  },
-  {
-    title: "Red Sparrow",
-    director: "Francis Lawrence",
-    stars: ["Jennifer Lawrence", "Joel Edgerton", "Matthias Schoenaerts"],
-    image:
-      "https://images-na.ssl-images-amazon.com/images/M/MV5BMTA3MDkxOTc4NDdeQTJeQWpwZ15BbWU4MDAxNzgyNTQz._V1_UX182_CR0,0,182,268_AL_.jpg",
-    description:
-      "Ballerina Dominika Egorova is recruited to 'Sparrow School,' a Russian intelligence service where she is forced to use her body as a weapon. Her first mission, targeting a C.I.A. agent, threatens to unravel the security of both nations.",
-    showtimes: ["13:00", "15:30", "18:00", "20:10", "22:40"]
+async function seedMovies() {
+  try {
+    await mongooseConnect();
+    await mongooseDropCollection();
+    const movies = await Movie.insertMany(moviesData);
+    console.log('movies', movies);
+    await mongooseClose();
+  } catch (error) {
+    console.error(`Error: ${error.message}`);
   }
-];
-  
+}
+
+seedMovies();
+
+// Before
+// // To insert in "seeds/movies.seed.js"
+// const mongoose = require("mongoose");
+// const Movie = require("../models/Movie.model");
+// const movies = require("./movies.data"); // require the movies from --> movies.data.js
+
+// const MONGO_URI = process.env.MONGODB_URI || "mongodb://localhost/express-cinema";
+
+// mongoose.connect(MONGO_URI, {
+//   useNewUrlParser: true,
+//   useUnifiedTopology: true,
+//   //useFindAndModify: false,
+//   //useCreateIndex: true
+// }).then(self => {
+//   console.log(`Connected to the database: "${self.connection.name}"`);
+//   console.log('Deleting DB')
+//   // Before adding any recipes to the database, let's remove all existing ones
+//   return Movie.deleteMany();
+// })
+// .then( async() => {
+//   try {
+//     // const deleteMovies =  mongoose.connection.dropCollection('movies');
+//     // console.log('Drop DB');
+//     const addMovies = await Movie.create(movies);
+//     console.log(`Created ${addMovies.length} movies`);
+//     mongoose.connection.close();
+//   } catch(error) {
+//     console.error('There was an error', error);
+//   }
+// }).catch(err => console.log(`An error occurs while creating the database ${err}`));
+
+
 // Add here the script that will be run to actually seed the database (feel free to refer to the previous lesson)
 // Create the Model
-Movie.create(movies)
-  .then(moviesFromDB => {
-    console.log(`Created ${moviesFromDB.length} movies`);
-    // Once created close the DB connection
-    mongoose.connection.close();
-  })
-  .catch(err => console.log(`An error occurs while creating the database ${err}`));
+
+// Movie.create(movies)
+//   .then(moviesFromDB => {
+//     console.log(`Created ${moviesFromDB.length} movies`);
+//     // Once created close the DB connection
+//     mongoose.connection.close();
+//   })
+//   .catch(err => console.log(`An error occurs while creating the database ${err}`));
