+++ views/movies/movie-detail.hbs
@@ -1,21 +1,21 @@
 <div class="details-wrapper">
   <div class="img-wrapper">
-    <img src={{movie.image}} alt="">
+    <img src={{image}} alt="">
   </div>
   <div class="main-info">
     <a href="/">Go back</a>
-    <h2>{{movie.title}}</h2>
-    <h3>Director: {{movie.director}}</h3>
+    <h2>{{title}}</h2>
+    <h3>Director: {{director}}</h3>
     <p>Stars: </p>
     <ul>
-      {{#each movie.stars}}
+      {{#each stars}}
         <li>{{this}}</li>
       {{/each}}
     </ul>
-    <p>{{movie.description}}</p>
+    <p>{{description}}</p>
     <div>
       <p>Showtimes:</p>
-      {{#each movie.showtimes}}
+      {{#each showtimes}}
         <span>{{this}} | </span>
       {{/each}}
     </div>
